First tests pass for numeric integration system.
I made a few core classes for dealing with allocating temporary buffers, which we do not do blindly using malloc because that hurts performance and safety for allocations with small scope, especially when in a thread.
I also rolled a fixed-size vector class which is guaranteed never to allocate memory and does all bounds checking in assert() statements that can be reduced to no-ops with a compile flag.
Then I made an RK integrator for explicit methods with arbitrary butcher tableau (default is rk4) and tested some coupled systems.

Next thing to do is add support for parameters and coupling topology, and then we'll move to parallel parameter sweeps and then add in the bridge stuff.